+++ stories/index.js
@@ -1,14 +1,14 @@
 import {storiesOf} from '@storybook/vue'
 import 'element-ui/lib/theme-chalk/index.css'
 import Basic from './basic.vue'
-import SearchDataTable from './search-data-table.vue'
+import Search from './search.vue'
 import TreeDataTable from './tree-data-table.vue'
 import CustomFormDataTable from './custom-form-data-table.vue'
 import CustomButtonDataTable from './custom-button-data-table.vue'
 
 storiesOf('ElDataTable', module)
   .add('基本CRUD', basic)
-  .add('查询表格', searchTable)
+  .add('基本查询', search)
   .add('树形表格', treeTable)
   .add('自定义表单', customFormTable)
   .add('自定义操作按钮', customButtonTable)
@@ -25,10 +25,10 @@ function treeTable() {
     template: `<tree-data-table/>`
   }
 }
-function searchTable() {
+function search() {
   return {
-    components: {SearchDataTable},
-    template: `<search-data-table></search-data-table>`
+    components: {Search},
+    template: `<search></search>`
   }
 }
 function customFormTable() {
