--- stories/search-data-table.vue
@@ -1,30 +1,0 @@
-<template>
-  <el-data-table
-      :url="url"
-      :columns="columns"
-      :form="form"
-      :searchForm="searchForm"
-      :formAttrs="formAttrs"
-      :dataPath="dataPath"
-      >
-      </el-data-table>
-</template>
-<script>
-import ElDataTable from '../src/el-data-table.vue'
-import config from './config'
-export default {
-  components: {ElDataTable},
-  data: function() {
-    return {
-      url: config.url,
-      dataPath: 'data',
-      form: config.form,
-      formAttrs: config.formAttrs,
-      searchForm: config.form.map(form => {
-        return Object.assign({}, form, {rules: []})
-      }),
-      columns: config.columns
-    }
-  }
-}
-</script>
